+++ resources/js/Utils/formatters.ts
@@ -12,3 +12,18 @@ export const formatDateTime = (dateString: string): string => {
 export const formatCurrency = (amount: number): string => {
   return `€${amount.toFixed(2)}`;
 };
+
+export const formatTime = (timeString: string) => {
+  const date = new Date(timeString);
+  return date.toLocaleTimeString('en-US', {
+    hour: '2-digit',
+    minute: '2-digit',
+  });
+};
+
+export const formatCurrencyUSD = (amount: number) => {
+  return new Intl.NumberFormat('en-US', {
+    style: 'currency',
+    currency: 'USD',
+  }).format(amount);
+};
